moved code out of index.js
