+++ pages/index.js
@@ -1,18 +1,18 @@
 import React, { useState, useEffect } from 'react';
 import {
-  Row, Col, Button, ButtonGroup, ButtonToolbar,
-  ListGroup, Offcanvas, ProgressBar, Alert
+  Row, Col, ButtonToolbar, ListGroup, Offcanvas,
+  ProgressBar, Alert
 } from 'react-bootstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCircleNotch, faExclamationCircle, faArrowsRotate } from '@fortawesome/free-solid-svg-icons';
+import { faCircleNotch } from '@fortawesome/free-solid-svg-icons';
 import { faLink } from '@fortawesome/free-solid-svg-icons';
 import { Layout } from '../components/Layout'
 import DatasetSelector from '../components/DatasetSelector';
 import LogsComponent from '../components/LogsComponent';
+import ErrorPagePopup from '../components/ErrorPagePopup';
 import MilestonesSidebar from '../components/MilestonesSidebar';
 import {
-  TaskCompleteCheckbox,
-  MainThreeActionButtons
+  TaskCompleteCheckbox, MainThreeActionButtons
 } from '../components/ActionButtons';
 import { makeUseAxios } from 'axios-hooks'
 import {
@@ -27,51 +27,13 @@ export default function Index(props) {
   const [showDebugData, setshowDebugData] = useState(false);
   const [workflow, setWorkflow] = useState();
 
-  const [{ loading, error }, makeApiRequest] = useAxios(
+  const [{ loading, error: apiError }, makeApiRequest] = useAxios(
     null, { manual: true }
   );
   async function fetchWorkflow() {
     makeApiRequest(
       getWorkflow(props.currentDatasetId)
     ).then(res => setWorkflow(res.data))
-  }
-  function ErrorPageComponent() {
-    const [showJsonDump, setshowJsonDump] = useState(false);
-    const errorLogs = [
-      { title: 'API Request', data: error },
-      { title: 'workflow', data: workflow },
-      { title: 'props', data: props }
-    ];
-    return (
-      <Offcanvas show={true} placement="top" keyboard={false}>
-        <Offcanvas.Body>
-          <div className="text-center">
-            <h4 className="text-danger">
-              <FontAwesomeIcon icon={faExclamationCircle} className="me-2" />
-              <span>Something went wrong, please refresh this page</span>
-            </h4>
-            <ButtonGroup size="sm" className="mt-2">
-              <Button
-                variant="danger"
-                onClick={() => location.reload(true)}
-              >
-                <FontAwesomeIcon icon={faArrowsRotate} className="me-2" />
-                <span>Refresh Page</span>
-              </Button>
-              <Button
-                variant="light"
-                onClick={() => setshowJsonDump(!showJsonDump)}
-              >View Error Logs</Button>
-            </ButtonGroup>
-          </div>
-          {showJsonDump && (
-            <Col xs={{ span: 6, offset: 3 }}>
-              <LogsComponent objects={errorLogs} />
-            </Col>
-          )}
-        </Offcanvas.Body>
-      </Offcanvas>
-    )
   }
 
   useEffect(() => {
@@ -251,7 +213,7 @@ export default function Index(props) {
           </h2>
         </Offcanvas.Body>
       </Offcanvas>
-      {error && <ErrorPageComponent />}
+      {apiError && <ErrorPagePopup {...{ apiError, workflow, props }} />}
       {showDebugData && (
         <LogsComponent objects={[
           { title: 'workflow', data: workflow },
